Removed init.php as file and from autoload in composer.json Changed psr-4 in composer.json
