+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "larakit/hlp-image",
-    "description": "[Larakit] [helper] file",
+    "description": "[Larakit] [helper] image",
     "keywords": [
         "laravel 5",
         "laravel",
@@ -14,16 +14,11 @@
         }
     ],
     "require": {
-        "larakit/lk-boot": "*",
         "intervention/image": "*"
     },
     "autoload": {
         "psr-4": {
-            "Larakit\\Helpers\\": "src\\Helpers"
-        },
-        "files": [
-            "src/init.php"
-        ]
-    },
-    "minimum-stability": "dev"
+            "Larakit\\Helpers\\": "src/"
+        }
+    }
 }
